+++ CHANGELOG.md
@@ -145,6 +145,7 @@
 - AddVersionListener and VersionBuilder use new `pim_versioning.serializer` service
 - In InGroupFilter and IsAssociatedFilter constructors, replace the RequestParameters argument by a RequestParametersExtractorInterface
 - Change constructor of `Pim\Bundle\DataGridBundle\Controller\ProductExportController` to inject the product repository `Pim\Bundle\CatalogBundle\Repository\ProductRepositoryInterface` as fourth argument
+- Rename Pim\CatalogBundle\Model\Media to Pim\CatalogBundle\Model\ProductMedia to reflect the coupling between product media and product value and allow the future introduction of generic media
 - Add a $metricClass parameter in MetricTransformer constructor
 
 # 1.1.0 - "Rabbit Punch" (2014-04-16)